Update on "[vulkan] Enable prepacked addmm for linear layers"
Differential Revision: [D24908605](https://our.internmc.facebook.com/intern/diff/D24908605)

[ghstack-poisoned]
