+++ aten/src/ATen/native/vulkan/ops/Convolution.h
@@ -1,0 +1,99 @@
+#pragma once
+#ifdef USE_VULKAN
+
+#include <ATen/native/vulkan/ops/Common.h>
+#include <torch/custom_class.h>
+
+namespace at {
+namespace native {
+namespace vulkan {
+namespace ops {
+
+class Conv2dOpContext final : public torch::jit::CustomClassHolder {
+ public:
+  static Conv2dOpContext create(
+      api::Resource::Pool& pool,
+      const Tensor& weight,
+      const c10::optional<Tensor>& bias,
+      IntArrayRef stride,
+      IntArrayRef padding,
+      IntArrayRef dilation,
+      bool transposed,
+      IntArrayRef output_padding,
+      int64_t groups,
+      c10::optional<Scalar> output_min = c10::nullopt,
+      c10::optional<Scalar> output_max = c10::nullopt);
+
+  using State = std::tuple<
+      Tensor,
+      c10::optional<Tensor>,
+      std::vector<int64_t>,
+      std::vector<int64_t>,
+      std::vector<int64_t>,
+      int64_t,
+      c10::optional<Scalar>,
+      c10::optional<Scalar>>;
+
+  Tensor run(const Tensor& input) const;
+  State unpack() const;
+
+ private:
+  Conv2dOpContext(
+      api::Resource::Pool& pool,
+      const Tensor& weight,
+      const c10::optional<Tensor>& bias,
+      IntArrayRef stride,
+      IntArrayRef padding,
+      IntArrayRef dilation,
+      bool transposed,
+      IntArrayRef output_padding,
+      int64_t groups,
+      c10::optional<Scalar> output_min = c10::nullopt,
+      c10::optional<Scalar> output_max = c10::nullopt);
+
+ private:
+  struct {
+    vTensor v_weight;
+    vTensor v_bias;
+    std::array<int64_t, 4> filter;
+    std::array<int64_t, 2> stride;
+    std::array<int64_t, 2> padding;
+    std::array<int64_t, 2> dilation;
+    int32_t groups;
+    float output_min;
+    float output_max;
+  } packed_;
+
+  struct {
+    Tensor weight;
+    c10::optional<Tensor> bias;
+    std::vector<int64_t> filter;
+    std::vector<int64_t> stride;
+    std::vector<int64_t> padding;
+    std::vector<int64_t> dilation;
+    int64_t groups;
+    c10::optional<Scalar> output_min;
+    c10::optional<Scalar> output_max;
+  } unpacked_;
+};
+
+Tensor conv2d_clamp_run(
+    const Tensor& input,
+    const c10::intrusive_ptr<Conv2dOpContext>& context);
+
+c10::intrusive_ptr<Conv2dOpContext> conv2d_clamp_prepack(
+    Tensor&& weight,
+    c10::optional<Tensor>&& bias,
+    std::vector<int64_t>&& stride,
+    std::vector<int64_t>&& padding,
+    std::vector<int64_t>&& dilation,
+    const int64_t groups,
+    const c10::optional<Scalar> output_min,
+    const c10::optional<Scalar> output_max);
+
+} // namespace ops
+} // namespace vulkan
+} // namespace native
+} // namespace at
+
+#endif /* USE_VULKAN */
