+++ aten/src/ATen/native/vulkan/ops/RegisterOpContextClass.cpp
@@ -1,0 +1,80 @@
+#ifdef USE_VULKAN
+
+#include <torch/library.h>
+#include <torch/custom_class.h>
+#include <ATen/native/vulkan/ops/Common.h>
+#include <ATen/native/vulkan/ops/Convolution.h>
+#include <ATen/native/vulkan/ops/Mm.h>
+
+namespace at {
+namespace native {
+namespace vulkan {
+namespace ops {
+namespace {
+
+TORCH_LIBRARY(vulkan, m) {
+  m.class_<Conv2dOpContext>("Conv2dOpContext")
+      .def_pickle(
+          // __getstate__
+          [](const c10::intrusive_ptr<Conv2dOpContext>& context) {
+            return context->unpack();
+          },
+          // __setstate__
+          [](Conv2dOpContext::State state) {
+            return conv2d_clamp_prepack(
+                std::move(std::get<0>(state)),
+                std::move(std::get<1>(state)),
+                std::move(std::get<2>(state)),
+                std::move(std::get<3>(state)),
+                std::move(std::get<4>(state)),
+                std::move(std::get<5>(state)),
+                std::move(std::get<6>(state)),
+                std::move(std::get<7>(state)));
+          });
+  m.class_<LinearOpContext>("LinearOpContext")
+      .def_pickle(
+          // __getstate__
+          [](const c10::intrusive_ptr<LinearOpContext>& context) {
+            return context->unpack();
+          },
+          // __setstate__
+          [](LinearOpContext::State state) {
+            return linear_prepack(
+                std::move(std::get<0>(state)), std::move(std::get<1>(state)));
+          });
+}
+
+TORCH_LIBRARY(vulkan_prepack, m) {
+  m.def(
+      "conv2d_clamp_prepack(Tensor W, Tensor? B, int[2] stride, "
+      "int[2] padding, int[2] dilation, int groups, "
+      "Scalar? output_min=None, Scalar? output_max=None) "
+      "-> __torch__.torch.classes.vulkan.Conv2dOpContext");
+  m.def(
+      "conv2d_clamp_run(Tensor X, "
+      "__torch__.torch.classes.vulkan.Conv2dOpContext W_prepack) -> Tensor Y");
+  m.def(
+      "linear_prepack(Tensor W, Tensor? B) "
+      "-> __torch__.torch.classes.vulkan.LinearOpContext");
+  m.def(
+      "linear_run(Tensor X, "
+      "__torch__.torch.classes.vulkan.LinearOpContext BW_prepack) -> Tensor Y");
+}
+
+TORCH_LIBRARY_IMPL(vulkan_prepack, CPU, m) {
+  m.impl("conv2d_clamp_prepack", TORCH_FN(conv2d_clamp_prepack));
+  m.impl("linear_prepack", TORCH_FN(linear_prepack));
+}
+
+TORCH_LIBRARY_IMPL(vulkan_prepack, Vulkan, m) {
+  m.impl("conv2d_clamp_run", TORCH_FN(conv2d_clamp_run));
+  m.impl("linear_run", TORCH_FN(linear_run));
+}
+
+} // namespace
+} // namespace ops
+} // namespace vulkan
+} // namespace native
+} // namespace at
+
+#endif /* USE_VULKAN */
